backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
playwright
python-dotenv
pydantic